requirements.txt
pandas
scikit-learn
streamlit
plotly
groq
python-dotenv

# NOTE: run python -m spacy download en_core_web_sm after installing all the libraries!!